requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.0,<5.0
django-environ>=0.10,<1.0
Pillow>=10.0,<11.0
gunicorn>=21.0,<22.0
django-crontab>=0.7.1



django-crispy-forms
django-widget-tweaks
django-filter
django-import-export
django-extensions

python-decouple
ipython
pytest-django
coverage
